new_src/requirements.txt
torch>=1.7.0
torchvision>=0.8.0
tensorboardX>=2.1
numpy>=1.19.0
Pillow>=8.0.0
tqdm>=4.50.0 